Add files via upload
"Recently I came across a new library for me, lifelines. I did a bit of exploration on how people use it and read a small article. I wanted to do a small simple analysis using this library and throw this notebook into my portfolio. We'll take a dataset from the Kaggle platform as a basis. The goal there was to predict behavior to retain customers. Honestly, I'm not an expert in forecasts right now, as it's a whole section of ML, although some logistic regression might be applicable here. Right now, I just want to do a basic analysis of the dataset, including the new library. I want to look at churn, or the so-called Churn Rate, and visualize this metric including various factors. [Example datasets IBM]. Link: https://www.kaggle.com/datasets/blastchar/telco-customer-churn/data"